+++ tests/unit/balance.spec.ts
@@ -1,10 +1,12 @@
 import { parseUnits } from "viem";
 import {
-  AssetTransfer,
   assetTransfersToSummary,
   checkAllBalances,
+} from "../../src/balance";
+import {
+  AssetTransfer,
   CollectibleTransfer,
-} from "../src";
+} from "../../src/csv"
 
 const unlistedERC20Token = {
   address: "0x6b175474e89094c44da98b954eedeac495271d0f",
